refactor: make the toolbar a registry
The toolbar had no array, only three anchored groups of markup. Extract
ModeSwitcher, PageTitlePopover, ViewerAvatars, ReadOnlyBadge and
ToolbarActions into components/ToolbarItems, then register them with a
region.

ToolbarActions stays one item. The settings, preview and AI icons share a
gap-2 group inside a gap-4 region, so five separate items would change the
spacing. The four mode buttons stay one item for the same reason.

MainMenu wrote its Settings and Shortcuts actions as emits. A registry item
has no parent to emit to, so it writes builderStore instead. Only the
toolbar renders MainMenu.

Delete the unused isDark, toggleDark and transitionTheme from the toolbar.
Only a commented-out button referenced them.

Part of extensibility phase 1, step 7.
