+++ frontend/src/components/MainMenu.vue
@@ -15,6 +15,7 @@
 </template>
 <script setup lang="ts">
 import { useDashboardState } from "@/composables/useDashboardState";
+import useBuilderStore from "@/stores/builderStore";
 import useCanvasStore from "@/stores/canvasStore";
 import usePageStore from "@/stores/pageStore";
 import { BuilderPage } from "@/types/doctypes";
@@ -31,8 +32,8 @@ const isDark = useDark({
 const router = useRouter();
 const toggleDark = useToggle(isDark);
 const canvasStore = useCanvasStore();
-
-const emit = defineEmits(["showSettings", "showShortcuts"]);
+// a registry item has no parent to emit to, so write the store directly
+const builderStore = useBuilderStore();
 
 const handleCopyPage = () => {
 	if (!pageStore.activePage) return;
@@ -91,8 +92,16 @@ const mainMenuOptions = [
 				onClick: () => toggleDark(),
 				icon: isDark ? "lucide-sun" : "lucide-moon",
 			},
-			{ label: "Settings", onClick: () => emit("showSettings"), icon: "lucide-settings" },
-			{ label: "Shortcuts", onClick: () => emit("showShortcuts"), icon: "lucide-command" },
+			{
+				label: "Settings",
+				onClick: () => (builderStore.showSettingsDialog = true),
+				icon: "lucide-settings",
+			},
+			{
+				label: "Shortcuts",
+				onClick: () => (builderStore.shortcutsModalOpen = true),
+				icon: "lucide-command",
+			},
 			{
 				label: "Help",
 				onClick: () => {
